New result report script.
